# specs/openapi.yaml
openapi: 3.0.3
info:
  title: FreqShow API
  version: 0.1.0
  description: |
    FreqShow backend API.

    This spec is intended to become the source of truth for backend responses
    and generated frontend types.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /auth/request:
    post:
      tags: [Auth]
      summary: Request a magic-link sign-in email
      description: |
        Accepts an email address and, if login is configured (RESEND_API_KEY
        set), emails a time-limited sign-in link. The response is the same
        whether or not the address already has an account — accounts are
        created lazily on first successful GET /auth/verify — so this
        endpoint cannot be used to enumerate registered emails.
      operationId: requestLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestBody'
      responses:
        '200':
          description: Sign-in email sent (or accepted for sending)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRequestResponse'
        '400':
          description: Invalid request body or malformed email address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login requests from this client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Login is not configured (no Resend API key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
components:
...
  responses:
    MethodNotAllowed:
      description: Method not allowed
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...
    AuthRequestBody:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
    AuthRequestResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
        message:
          type: string
...